+++ src/prelude.rs
@@ -352,6 +352,10 @@ pub unsafe trait WindowExt: GroupExt {
     /// # Safety
     /// The data must be valid.
     unsafe fn set_region(&mut self, region: crate::draw::Region);
+    /// Iconifies the window
+    fn iconize(&mut self);
+    /// Returns whether the window is fullscreen or not
+    fn fullscreen_active(&self) -> bool;
 }
 
 /// Defines the methods implemented by all input and output widgets